Some progress on escrow invariant; a big challenge: scoping; e.g. want
to get a points-to out of an invariant in a block of code, and to the
outer block of code. To be sure that a predeclared variable is
initialized, want to rely on one branch being unreachable, but can't
figure out a way to communicate that to the type checker.